base function editor layout and response to open-source-event
