+++ src/modules/project_explorer/init_project_explorer.c
@@ -14,8 +14,6 @@ int init_project_explorer(mc_node *app_root)
   //   MCcall(mcs_interpret_file("src/modules/source_editor/function_editor.c"));
   //   MCcall(mcs_interpret_file("src/modules/source_editor/source_editor_pool.c"));
 
-  //   MCitpCall(mce_init_source_editor_pool, "mce_init_source_editor_pool");
-
   mc_app_itp_data *itp_data;
   mc_obtain_app_itp_data(&itp_data);
   int (*mcm_init_project_explorer)(mc_node *) = tcci_get_symbol(itp_data->interpreter, "mcm_init_project_explorer");
