+++ src/modules/source_editor/init_source_editor.c
@@ -20,13 +20,12 @@
 int init_source_editor(mc_node *app_root)
 {
   MCcall(mcs_interpret_file("src/modules/source_editor/source_editor.h"));
-  MCcall(mcs_interpret_file("src/modules/source_editor/source_line.c"));
-  MCcall(mcs_interpret_file("src/modules/source_editor/function_editor.c"));
-  MCcall(mcs_interpret_file("src/modules/source_editor/source_editor_pool.c"));
+  MCcall(mcs_interpret_file("src/modules/source_editor/source_editor.c"));
+  // MCcall(mcs_interpret_file("src/modules/source_editor/source_line.c"));
+  // MCcall(mcs_interpret_file("src/modules/source_editor/function_editor.c"));
+  // MCcall(mcs_interpret_file("src/modules/source_editor/source_editor_pool.c"));
 
-  MCitpCall(mce_init_source_editor_pool, "mce_init_source_editor_pool");
-
-  // MCcall(mce_init_source_editor_pool());
+  MCitpCall(mc_se_init_source_editor, "mc_se_init_source_editor");
 
   return 0;
 }
